+++ pkg/cloudwrapper/cloudwrapper.go
@@ -15,6 +15,7 @@ var (
 type (
 	// CloudWrapper is the api interface for Cloud Wrapper
 	CloudWrapper interface {
+		Capacities
 	}
 
 	cloudwrapper struct {