+++ pkg/cloudwrapper/mocks.go
@@ -3,6 +3,8 @@
 package cloudwrapper
 
 import (
+	"context"
+
 	"github.com/stretchr/testify/mock"
 )
 
@@ -11,3 +13,12 @@ type Mock struct {
 }
 
 var _ CloudWrapper = &Mock{}
+
+// ListCapacities implements CloudWrapper
+func (m *Mock) ListCapacities(ctx context.Context, req ListCapacitiesRequest) (*ListCapacitiesResponse, error) {
+	args := m.Called(ctx, req)
+	if args.Get(0) == nil {
+		return nil, args.Error(1)
+	}
+	return args.Get(0).(*ListCapacitiesResponse), args.Error(1)
+}